AudioRecorder 객체 생성
